SSP-490. Undelete user on receiving early alert.
Patch is broader than you might expect because it refactors similar code for handling
program status initialization on early alert receipt as well. Previously that code
was located in the PearsonEarlyAlertController, but both undeleting and setting
program statuses seemed like such core "business logic" that moving it into the
service next to the email alerting code seemed to make sense.